# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: QuizzGame API — Practice, Thi thử, Admin & Ngân hàng câu hỏi
  version: "1.3.0"
  description: >-
    Contract cho module Ôn tập (Practice), module Thi thử (Exam/Mock Exam),
    các endpoint Admin liên quan (quản lý câu hỏi, xử lý báo cáo câu hỏi,
    quản lý đề thi thử) và Ngân hàng câu hỏi (Question Bank).
    Khớp với `docs/FEATURE_LOG.md` Section 4 (Practice Module), Section 5
    (Admin Dashboard – Quản lý báo cáo câu hỏi), Section 6 (Exam Module –
    Thi thử) và Section 7 (Ngân hàng câu hỏi).

    Các endpoint khác của hệ thống (Auth, Onboarding, Points, Leaderboard...)
    chưa được mô tả trong file này.
servers:
  - url: http://localhost:4000
    description: Local dev
...
security: []
paths:
...
  /api/admin/exam-papers:
    post:
      tags: [Admin]
      summary: Tạo đề thi thử mới (chưa có câu hỏi)
      security: [{ AdminSecret: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExamPaperInput'
            example:
              subject: TOAN
              title: 'Đề thi thử THPT QG 2024 - Mã đề 101'
              durationMinutes: 50
      responses:
        '201':
          description: Đề thi đã được tạo (chưa có câu hỏi)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamPaperSummaryDto'
              example:
                id: b2c3d4e5-0001-4a3a-9b6e-9b6c1a2b3c4d
                subject: TOAN
                title: 'Đề thi thử THPT QG 2024 - Mã đề 101'
                durationMinutes: 50
                isActive: true
                questionCount: 0
                createdAt: '2026-06-15T08:00:00.000Z'
        '400':
          description: '`subject` không thuộc `SUBJECT_CATALOG`, `title` rỗng/>300, hoặc `durationMinutes` không hợp lệ (Zod)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
    get:
      tags: [Admin]
      summary: Danh sách đề thi thử (lọc theo môn)
      description: >-
        ⚠️ Trả về **mảng trực tiếp**, KHÔNG phân trang, KHÔNG bọc `items`
        (khác với `GET /api/admin/questions`).
      security: [{ AdminSecret: [] }]
      parameters:
        - name: subject
          in: query
          required: false
          description: Nếu không truyền, trả tất cả đề (mọi môn)
          schema:
            $ref: '#/components/schemas/SubjectId'
      responses:
        '200':
          description: 'Danh sách đề thi (`questionCount` chỉ đếm câu `isActive=true`)'
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExamPaperSummaryDto' }
              example:
                - id: b2c3d4e5-0001-4a3a-9b6e-9b6c1a2b3c4d
                  subject: TOAN
                  title: 'Đề thi thử THPT QG 2024 - Mã đề 101'
                  durationMinutes: 50
                  isActive: true
                  questionCount: 3
                  createdAt: '2026-06-15T08:00:00.000Z'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
...
components:
  securitySchemes:
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token nội bộ nhận được sau khi đăng nhập Firebase (xem Section 3 - Auth)
    AdminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Giá trị biến môi trường `ADMIN_SECRET`
...
  responses:
    AdminUnauthorized:
      description: Thiếu hoặc sai `X-Admin-Secret`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: ADMIN_UNAUTHORIZED
            message: 'Khong co quyen truy cap.'
  schemas:
    SubjectId:
      type: string
      enum: [TOAN, VAN, ANH, LY, HOA, SINH, SU, DIA, GDCD]
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Mã lỗi (dùng để ánh xạ HTTP status, xem `app.ts`)
        message:
          type: string
        details:
          type: array
          description: Chỉ có khi `error = INVALID_REQUEST_BODY` (lỗi Zod)
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
...
    ExamPaperSummaryDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        subject: { $ref: '#/components/schemas/SubjectId' }
        title: { type: string }
        durationMinutes: { type: integer }
        isActive: { type: boolean }
        questionCount:
          type: integer
          description: Số câu hỏi ĐANG KÍCH HOẠT (isActive=true) trong đề
        createdAt: { type: string, format: date-time }
...
    CreateExamPaperInput:
      type: object
      required: [subject, title, durationMinutes]
      properties:
        subject: { $ref: '#/components/schemas/SubjectId' }
        title: { type: string, minLength: 1, maxLength: 300 }
        durationMinutes: { type: integer, minimum: 1, maximum: 600 }
...